+++ app/controllers/additionals_macros_controller.rb
@@ -1,0 +1,7 @@
+class AdditionalsMacrosController < ApplicationController
+  before_action :require_login
+
+  def show
+    @available_macros = AdditionalsMacro.all
+  end
+end