+++ app/helpers/additionals_queries_helper.rb
@@ -12,13 +12,6 @@ module AdditionalsQueriesHelper
           params[:set_filter] ||
           session[session_key].nil? ||
           session[session_key][:project_id] != (@project ? @project.id : nil)
-
-      if options[:with_default_query] && !api_request? && %i[op f fields].all? { |k| !params.key?(k) }
-        d_query = query_class.default_query
-        if d_query.present? && query_class.where(id: d_query.id).exists?
-          return additionals_load_query_id(query_class, session_key, d_query.id, options, object_type)
-        end
-      end
       # Give it a name, required to be valid
       @query = query_class.new(name: '_')
       @query.project = @project
@@ -93,7 +86,7 @@ module AdditionalsQueriesHelper
     scope = scope.order(last_login_on: :desc)
                  .limit(params[:limit] || Additionals::SELECT2_INIT_ENTRIES)
     @users = scope.to_a.sort! { |x, y| x.name <=> y.name }
-    render layout: false, partial: 'auto_completes/users'
+    render layout: false, partial: 'auto_completes/additionals_users'
   end
 
   def additionals_query_to_xlsx(items, query, options = {})