+++ app/overrides/roles/form.rb
@@ -1,0 +1,5 @@
+Deface::Override.new virtual_path: 'roles/_form',
+                     name: 'roles-form-hide',
+                     insert_before: 'p.manage_members_shown',
+                     original: 'b2a317f49e0b65ae506c8871f0c2bcc3e8098766',
+                     partial: 'roles/additionals_form'