+++ lib/additionals/patches/role_patch.rb
@@ -1,0 +1,11 @@
+module Additionals
+  module Patches
+    module RolePatch
+      def self.included(base)
+        base.class_eval do
+          safe_attributes 'hide'
+        end
+      end
+    end
+  end
+end