+++ lib/additionals/patches/user_preference_patch.rb
@@ -3,7 +3,7 @@ module Additionals
     module UserPreferencePatch
       def self.included(base)
         base.class_eval do
-          safe_attributes 'autowatch_involved_issue' if Redmine::VERSION.to_s >= '3.4'
+          safe_attributes 'autowatch_involved_issue'
         end
       end
     end