+++ lib/additionals/wiki_macros/project_macro.rb
@@ -24,7 +24,7 @@ module Additionals
       macro :projects do |_obj, args|
         _args, options = extract_macro_options(args, :title, :with_create_issue)
         @projects = Additionals.load_projects
-        return '' if @projects.nil?
+        return if @projects.nil?
 
         @html_options = { class: 'external' }
         render partial: 'wiki/project_macros',