+++ test/unit/additionals_fontawesome_test.rb
@@ -18,7 +18,7 @@ class AdditionalsFontAwesomeTest < Additionals::TestCase
 
   def test_value_info_should_return_infos
     info = AdditionalsFontAwesome.value_info('invalid')
-    assert_equal [], info
+    assert_equal({}, info)
 
     info = AdditionalsFontAwesome.value_info('fas_car')
     assert info.key?(:classes)